arreglo merge

+++ Proyecto.02/project/src/RPGLogic.h
@@ -142,7 +142,7 @@ namespace rpgLogic {
 			_weaknesses[type] += change;
 		}
 
-		size_t getWeakness(damageType type) {	// Devolver la resistencia en función de tipo o parametro
+		float getWeakness(damageType type) {	// Devolver la resistencia en función de tipo o parametro
 			return _weaknesses[type];
 		}
 	};
